merge: モンスターノックバック処理を move_monster_to() に統一（hengband#5322 相当）
変愚蛮怒 PR hengband/hengband#5322「[Refactor] 烈風剣などモンスター
ノックバック処理の統一」を bakabakaband に適応してマージ。

従来 earthquake.cpp に private な static 関数として定義されていた
move_monster_to() を monster-util に昇格させ、烈風剣・衝波・狙撃など
モンスターノックバックを生じさせる複数箇所で共用できるようにした。

バグ修正効果: 烈風剣、衝波で視界外にモンスターを押し出したときに
反撃召喚などの update_monster() が行われなかったバグが修正される。

変更内容:
- monster-util.h: move_monster_to() を宣言
- monster-util.cpp: 実装を移動、grid.h / monster-update.h を追加 include
- earthquake.cpp: 重複実装を削除、monster-util.h を include
- combat/shoot.cpp: SP_RUSH 時のインライン実装を move_monster_to() 呼び出しに置換
- mind/mind-force-trainer.cpp: shock_power() を move_monster_to() に置換
- realm/realm-hissatsu.cpp: 烈風剣を move_monster_to() に置換

bakabakaband 構造への適応:
- PlayerType * → CreatureEntity &
- MonsterEntity & → CreatureEntity &
- player_ptr->current_floor_ptr → creature.get_floor()

上流コミット: 922cbe21c (hengband/develop)
関連 ISSUE: #7824
