+++ src/mind/mind-force-trainer.cpp
@@ -16,6 +16,7 @@
 #include "monster/monster-describer.h"
 #include "monster/monster-status.h"
 #include "monster/monster-update.h"
+#include "monster/monster-util.h"
 #include "pet/pet-util.h"
 #include "player-base/player-class.h"
 #include "player-info/equipment-info.h"
@@ -242,14 +243,8 @@ bool shock_power(CreatureEntity &creature)
     }
 
     msg_format(_("%sを吹き飛ばした！", "You blow %s away!"), m_name.data());
-    floor.get_grid(pos_origin).m_idx = 0;
-    floor.get_grid(pos_target).m_idx = m_idx;
-    monster.y = pos_target.y;
-    monster.x = pos_target.x;
-
-    update_monster(creature, m_idx, true);
-    lite_spot(creature, pos_origin);
-    lite_spot(creature, pos_target);
+    monster.set_target(creature.get_position());
+    move_monster_to(creature, monster, pos_target);
 
     if (monrace.brightness_flags.has_any_of(ld_mask)) {
         RedrawingFlagsUpdater::get_instance().set_flag(StatusRecalculatingFlag::MONSTER_LITE);
